# docs/open-api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "DiffScope Synthesis Platform API",
    "version": "0.0.0",
    "description": "DSSP HTTP API surface represented by this service. A server implementation can support a subset of the broader DSSP API."
  },
  "servers": [
    {
      "url": "http://localhost:13711",
      "description": "Default local server URL. The actual host and port are configured by the service."
    }
  ],
...
  "paths": {
...
    "/api/arch/{arch_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitecture",
        "summary": "Get architecture metadata",
        "description": "Returns metadata for one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Architecture"
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ArchitectureOperationError"
          }
        }
      }
    },
...
    "/api/synth/pronunciation": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthPronunciation",
        "summary": "Generate pronunciations",
        "description": "Generates pronunciation candidates for lyric input. Set stream to true to request newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/PronunciationRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PronunciationComplete"
          },
          "204": {
            "$ref": "#/components/responses/EnvTagMatched"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/PronunciationOperationError"
          }
        }
      }
    },
    "/api/synth/phoneme": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthPhoneme",
        "summary": "Generate phonemes",
        "description": "Generates phoneme notes from pronunciation notes. Set stream to true to request newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/PhonemeRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PhonemeComplete"
          },
          "204": {
            "$ref": "#/components/responses/EnvTagMatched"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/PhonemeOperationError"
          }
        }
      }
    },
...
    "/api/synth/parameter": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthParameter",
        "summary": "Generate parameter values",
        "description": "Generates requested parameter values. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/ParameterRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ParameterResult"
          },
          "204": {
            "$ref": "#/components/responses/EnvTagMatched"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ParameterOperationError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ArchID": {
        "name": "arch_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Architecture identifier."
      },
...
      "DisplayLanguage": {
        "name": "display_language",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Preferred language for display metadata."
      }
    },
    "requestBodies": {
      "PronunciationRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PronunciationRequest"
            }
          }
        }
      },
      "PhonemeRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PhonemeRequest"
            }
          }
        }
      },
...
      "ParameterRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ParameterRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Architecture": {
        "description": "Architecture metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadata"
            }
          }
        }
      },
      "ArchitectureNotFound": {
        "description": "The requested architecture does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureNotFoundAPIError"
            }
          }
        }
      },
      "ArchitectureOperationError": {
        "description": "The architecture metadata request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureOperationAPIError"
            }
          }
        }
      },
...
      "PronunciationComplete": {
        "description": "Pronunciation generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON events; an error event can appear after the HTTP status has already been sent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PronunciationResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/PronunciationStreamEvent"
            }
          }
        }
      },
      "PhonemeComplete": {
        "description": "Phoneme generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON events; an error event can appear after the HTTP status has already been sent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PhonemeResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/PhonemeStreamEvent"
            }
          }
        }
      },
...
      "ParameterResult": {
        "description": "Parameter generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON events; an error event can appear after the HTTP status has already been sent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ParameterResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/ParameterStreamEvent"
            }
          }
        }
      },
...
      "EnvTagMatched": {
        "description": "The supplied environment tag matches the current environment tag, so no response body is returned."
      },
      "BadRequest": {
        "description": "The request body is not valid JSON, contains more than one JSON value, or fails request validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestError"
            }
          }
        }
      },
      "PronunciationOperationError": {
        "description": "The pronunciation request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PronunciationAPIError"
            }
          }
        }
      },
      "PhonemeOperationError": {
        "description": "The phoneme request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PhonemeAPIError"
            }
          }
        }
      },
...
      "ParameterOperationError": {
        "description": "The parameter request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ParameterAPIError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ArchitectureMetadata": {
        "type": "object",
        "required": [
          "id",
          "name",
          "pronunciation_mode",
          "phoneme_mode",
          "parameters",
          "audio_dependencies"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "name": {
            "type": "string"
          },
          "pronunciation_mode": {
            "$ref": "#/components/schemas/PronunciationMode"
          },
          "phoneme_mode": {
            "$ref": "#/components/schemas/PhonemeMode"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ArchitectureParameterMetadata"
            }
          },
          "audio_dependencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "PronunciationMode": {
        "type": "string",
        "enum": [
          "FULL",
          "SKIP"
        ]
      },
      "PhonemeMode": {
        "type": "string",
        "enum": [
          "FULL",
          "TOKEN_ONLY",
          "SKIP"
        ]
      },
      "ArchitectureParameterMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DirectArchitectureParameterMetadata"
          },
          {
            "$ref": "#/components/schemas/IndirectArchitectureParameterMetadata"
          }
        ]
      },
...
      "PronunciationRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/SingleSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/PronunciationInput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/EnvTag"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        },
        "additionalProperties": false
      },
      "PhonemeRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/SingleSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/PhonemeInput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/EnvTag"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        },
        "additionalProperties": false
      },
...
      "ParameterRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/ParameterInput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/EnvTag"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        },
        "additionalProperties": false
      },
...
      "SingleSingerContext": {
        "type": "object",
        "required": [
          "arch",
          "arch_extra",
          "singer"
        ],
        "properties": {
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "arch_extra": {
            "description": "Architecture-specific JSON payload."
          },
          "singer": {
            "$ref": "#/components/schemas/Singer"
          }
        },
        "additionalProperties": false
      },
...
      "EnvTag": {
        "type": "string",
        "description": "Previously returned environment tag. If it matches the current environment tag, the service can return 204."
      },
      "ResponseEnvTag": {
        "type": "string",
        "description": "Environment tag for client-side cache validation. If this field is omitted or empty, the client should not cache the response."
      },
      "PronunciationInput": {
        "type": "object",
        "required": [
          "notes"
        ],
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lyric"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PronunciationResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/PronunciationOutput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/ResponseEnvTag"
          }
        },
        "additionalProperties": false
      },
      "PronunciationOutput": {
        "type": "object",
        "required": [
          "notes"
        ],
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pronunciation"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PhonemeResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/PhonemeOutput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/ResponseEnvTag"
          }
        },
        "additionalProperties": false
      },
...
      "ParameterResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/ParameterOutput"
          },
          "env_tag": {
            "$ref": "#/components/schemas/ResponseEnvTag"
          }
        },
        "additionalProperties": false
      },
...
      "PronunciationStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/PronunciationResponse"
          },
          {
            "$ref": "#/components/schemas/PronunciationAPIError"
          }
        ]
      },
      "PhonemeStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/PhonemeResponse"
          },
          {
            "$ref": "#/components/schemas/PhonemeAPIError"
          }
        ]
      },
...
      "ParameterStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/ParameterProcessingEvent"
          },
          {
            "$ref": "#/components/schemas/ParameterResponse"
          },
          {
            "$ref": "#/components/schemas/ParameterAPIError"
          }
        ]
      },
...
      "ArchitectureNotFoundAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureNotFoundErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ArchitectureOperationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureOperationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "BadRequestError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PronunciationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/PronunciationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PhonemeAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/PhonemeErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ParameterAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ParameterErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ArchitectureNotFoundErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN_ARCH"
        ]
      },
      "ArchitectureOperationErrorCode": {
        "type": "string",
        "enum": [
          "INTERNAL_ERROR"
        ]
      },
...
    }
  }
}